unsorted/silver-umbrella/silver_umbrella/requirements/base.txt
django 
python-dotenv 
djangorestframework 
django-cors-headers 
channels 
channels-redis 
pillow 
cuid 
django-crispy-forms 
django-pylibmc 
djoser 
hypercorn 
uvicorn 
gunicorn 
aiohttp 
whitenoise 
coverage[toml]